refactor(rpc): move token handlers to routes/tokens.rs (backlog #12 phase 2a)
Peel the 8 SRC-20 handlers out of routes/mod.rs into their own
module. api_err goes pub(super) so tokens.rs (and the next
phase-2 slices) can share it.

Zero route-path change, zero behaviour change. Keeping each slice
small so follow-up PRs review in minutes.
